src/api-service/requirements.txt
fastapi>=0.115.9
httpx>=0.28.1
uvicorn>=0.34.0

grpcio>=1.71.0
grpcio-tools>=1.62.3